senses/sensor_hub.py
import subprocess

try:
    import sounddevice as sd
    import numpy as np
except Exception:
    sd = None
    np = None

# 👁️ Sensor Pseudocode Module

def scan_environment():
    # Placeholder for Bluetooth, WiFi, Audio, etc.
    print("🔍 Scanning environment...")
    return {
        "bt_devices": [],
        "wifi_networks": [],
        "audio_signals": []
    }

def mic_listen(duration=2, samplerate=44100):
    if sd is None or np is None:
        return "sounddevice or numpy not installed"
    print("🎧 Listening via mic...")
    recording = sd.rec(int(duration * samplerate), samplerate=samplerate, channels=1)
    sd.wait()
    volume = np.linalg.norm(recording)
    return f"Mic volume level: {round(volume, 3)}"

def scan_bluetooth():
    print("📡 Scanning Bluetooth...")
    output = subprocess.getoutput("bluetoothctl devices")
    return output.splitlines()

def read_esp32(port='/dev/ttyUSB0', baud=115200):
    try:
        import serial
    except Exception as e:
        return f"pyserial not installed: {e}"
    try:
        with serial.Serial(port, baud, timeout=2) as ser:
            line = ser.readline().decode().strip()
            return f"ESP32 says: {line}"
    except Exception as e:
        return f"Error reading ESP32: {e}"
