requirements.txt
sounddevice
numpy
pyserial
Flask
